se actualizo el template a admindlte-3

+++ database/migrations/2020_06_25_040604_create_usuario_rol_table.php
@@ -14,8 +14,8 @@ class CreateUsuarioRolTable extends Migration
     public function up()
     {
         Schema::create('usuario_rol', function (Blueprint $table) {
-            $table->unsignedInteger('rol_id');
-            $table->unsignedInteger('usuario_id');
+            $table->unsignedBigInteger('rol_id');
+            $table->unsignedBigInteger('usuario_id');
             $table->foreign('rol_id', 'fk_usuariorol_rol')->references('id')->on('rol')->onDelete('restrict')->onUpdate('restrict');
             $table->foreign('usuario_id', 'fk_usuariorol_usuario')->references('id')->on('usuario')->onDelete('restrict')->onUpdate('restrict');
             $table->charset = 'utf8mb4';
