+++ database/migrations/2020_06_25_041756_create_permiso_rol_table.php
@@ -14,8 +14,8 @@ class CreatePermisoRolTable extends Migration
     public function up()
     {
         Schema::create('permiso_rol', function (Blueprint $table) {
-            $table->unsignedInteger('rol_id');
-            $table->unsignedInteger('permiso_id');
+            $table->unsignedBigInteger('rol_id');
+            $table->unsignedBigInteger('permiso_id');
             $table->foreign('rol_id', 'fk_permisorol_rol')->references('id')->on('rol')->onDelete('restrict')->onUpdate('restrict');
             $table->foreign('permiso_id', 'fk_permisorol_permiso')->references('id')->on('permiso')->onDelete('restrict')->onUpdate('restrict');
             $table->charset = 'utf8mb4';
