+++ database/migrations/2020_06_25_043202_create_libro_prestamo_table.php
@@ -14,9 +14,9 @@ class CreateLibroPrestamoTable extends Migration
     public function up()
     {
         Schema::create('libro_prestamo', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('usuario_id');
-            $table->unsignedInteger('libro_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('usuario_id');
+            $table->unsignedBigInteger('libro_id');
             $table->date('fecha_prestamo');
             $table->string('prestado_a', 100);
             $table->date('fecha_devolucion')->nullable();
