+++ database/migrations/2020_06_28_234334_crear_tabla_menu_rol.php
@@ -14,8 +14,8 @@ class CrearTablaMenuRol extends Migration
     public function up()
     {
         Schema::create('menu_rol', function (Blueprint $table) {
-            $table->unsignedInteger('rol_id');
-            $table->unsignedInteger('menu_id');
+            $table->unsignedBigInteger('rol_id');
+            $table->unsignedBigInteger('menu_id');
             $table->foreign('rol_id', 'fk_menurol_rol')->references('id')->on('rol')->onDelete('restrict')->onUpdate('restrict');
             $table->foreign('menu_id', 'fk_menurol_menu')->references('id')->on('menu')->onDelete('cascade')->onUpdate('restrict');
             $table->charset = 'utf8mb4';
